+++ js/main.js
@@ -6,10 +6,9 @@ CONFIGURAÇÃO
 **************************************************************************************************************
 */
 
-import keyboardListener from './keyboardListener.js'
-
-const screen = document.getElementById('screen')
-const context = screen.getContext('2d')
+import createGame from "./createGame.js"
+import createKeyboardListener from './keyboardListener.js'
+import renderScreen from "./renderScreen.js"
 
 /*
 **************************************************************************************************************
@@ -17,110 +16,8 @@ GAME
 **************************************************************************************************************
 */
 
-const createGame = () => {
-    const state = {
-        players: {},
-        fruits: {}
-    }
-
-    const addPlayer = (command) => {
-        const playerId = command.playerId
-        const playerX = command.playerX
-        const playerY = command.playerY
-
-        state.players[playerId] = {
-            x: playerX,
-            y: playerY
-        }
-    }
-
-    const removePlayer = (command) => {
-        const playerId = command.playerId
-
-        delete state.players[playerId]
-    }
-
-    const addFruit = (command) => {
-        const fruitId = command.fruitId
-        const fruitX = command.fruitX
-        const fruitY = command.fruitY
-
-        state.fruits[fruitId] = {
-            x: fruitX,
-            y: fruitY
-        }
-    }
-
-    const removeFruit = (command) => {
-        const fruitId = command.fruitId
-
-        delete state.fruits[fruitId]
-    }
-
-    const movePlayer = (command) => {
-        const keyPressed = command.keyPressed
-        const playerId = command.playerId
-        const player = state.players[command.playerId]
-
-        const acceptedMovements = {
-            ArrowUp: function (player) {
-                if (player.y - 1 >= 0) {
-                    player.y--
-                }
-            },
-            ArrowRight: function (player) {
-                if (player.x + 1 < screen.width) {
-                    player.x++
-                }
-            },
-            ArrowDown: function (player) {
-                if (player.y + 1 < screen.height) {
-                    player.y++
-                }
-            },
-            ArrowLeft: function (player) {
-                if (player.x - 1 >= 0) {
-                    player.x--
-                }
-            }
-        }
-
-        const moveFunction = acceptedMovements[keyPressed]
-
-        if (player && moveFunction) {
-            moveFunction(player)
-            checkForFruitCollision(playerId)
-        }
-
-        return
-    }
-
-    const checkForFruitCollision = (playerId) => {
-        const player = state.players[playerId]
-
-        for (const fruitId in state.fruits) {
-            const fruit = state.fruits[fruitId]
-            console.log(`Checking ${playerId} and ${fruitId}`)
-
-            if (player.x === fruit.x && player.y === fruit.y) {
-                console.log(`Collision between ${playerId} and ${fruitId}`)
-                removeFruit({ fruitId: fruitId })
-            }
-        }
-
-    }
-
-    return {
-        movePlayer,
-        addPlayer,
-        removePlayer,
-        addFruit,
-        removeFruit,
-        state
-    }
-}
-
-const game = createGame()
+const screen = document.getElementById('screen')
+const game = createGame
 
 /*
 **************************************************************************************************************
@@ -128,30 +25,18 @@ USER INPUTS
 **************************************************************************************************************
 */
 
-const pressedKeyListener = keyboardListener
-pressedKeyListener.subscribe(game.movePlayer)
+const keyboardListener = createKeyboardListener(document)
+keyboardListener.subscribe(game.movePlayer)
 
 /*
 **************************************************************************************************************
-APRESENTAÇÂO
+VISUAL LAYOUT       
 **************************************************************************************************************
 */
 
-const renderScreen = () => {
-    context.fillStyle = '#fff'
-    context.clearRect(0, 0, 10, 10)
+renderScreen(screen, game, requestAnimationFrame)
 
-    for (const playerId in game.state.players) {
-        const player = game.state.players[playerId]
-        context.fillStyle = "#262626"
-        context.fillRect(player.x, player.y, 1, 1)
-    }
-    for (const fruitId in game.state.fruits) {
-        const fruit = game.state.fruits[fruitId]
-        context.fillStyle = 'green'
-        context.fillRect(fruit.x, fruit.y, 1, 1)
-    }
-    requestAnimationFrame(renderScreen)
-}
-
-renderScreen()
+game.addPlayer({ playerId: 'player1', playerX: 1, playerY: 1 })
+game.addPlayer({ playerId: 'player2', playerX: 4, playerY: 6 })
+game.addFruit({ fruitId: 'fruit1', fruitX: 5, fruitY: 8 })
+game.addFruit({ fruitId: 'fruit2', fruitX: 2, fruitY: 9 })
